# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 2gether Programming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomRequest {
  roomId?: string;
}

export interface Room {
  roomId: string;
  createdAt: string;
  userCount: number;
}

export interface ErrorResponse {
  error: string;
}