# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 2gether Programming API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  roomId: string;
  createdAt: string;
  userCount: number;
}